backend/requirements.txt
Flask>=3.0,<4
Flask-SQLAlchemy>=3.1,<4
Flask-Migrate>=4.0,<5
Flask-JWT-Extended>=4.6,<5
Flask-Bcrypt>=1.0,<2
Flask-Cors>=4.0,<5
psycopg[binary]>=3.1,<4
python-dotenv>=1.0,<2
SQLAlchemy>=2.0.36,<3
openpyxl>=3.1,<4